+++ BALSAMIC/commands/config/case.py
@@ -1,3 +1,4 @@
+"""Balsamic config case CLI."""
 import json
 import logging
 import os
@@ -5,10 +6,38 @@ from pathlib import Path
 
 import click
 
-from BALSAMIC import __version__ as balsamic_version
-from BALSAMIC.constants.analysis import BIOINFO_TOOL_ENV, Gender
+from BALSAMIC.commands.options import (
+    OPTION_GENOME_VERSION,
+    OPTION_ADAPTER_TRIM,
+    OPTION_ANALYSIS_DIR,
+    OPTION_ANALYSIS_WORKFLOW,
+    OPTION_BACKGROUND_VARIANTS,
+    OPTION_BALSAMIC_CACHE,
+    OPTION_CADD_ANNOTATIONS,
+    OPTION_CANCER_GERMLINE_SNV_OBSERVATIONS,
+    OPTION_CANCER_SOMATIC_SNV_OBSERVATIONS,
+    OPTION_CANCER_SOMATIC_SV_OBSERVATIONS,
+    OPTION_CASE_ID,
+    OPTION_CLINICAL_SNV_OBSERVATIONS,
+    OPTION_CLINICAL_SV_OBSERVATIONS,
+    OPTION_CONTAINER_VERSION,
+    OPTION_FASTQ_PATH,
+    OPTION_GENDER,
+    OPTION_NORMAL_SAMPLE_NAME,
+    OPTION_PANEL_BED,
+    OPTION_PON_CNN,
+    OPTION_QUALITY_TRIM,
+    OPTION_SWEGEN_SNV,
+    OPTION_SWEGEN_SV,
+    OPTION_TUMOR_SAMPLE_NAME,
+    OPTION_UMI,
+    OPTION_UMI_TRIM_LENGTH,
+)
+from BALSAMIC.constants.analysis import BIOINFO_TOOL_ENV, Gender, AnalysisWorkflow
+from BALSAMIC.constants.cache import GenomeVersion
 from BALSAMIC.constants.paths import CONTAINERS_DIR
 from BALSAMIC.constants.workflow_params import VCF_DICT
+from BALSAMIC.models.analysis import ConfigModel
 from BALSAMIC.utils.cli import (
     get_sample_list,
     get_panel_chrom,
@@ -17,204 +46,64 @@ from BALSAMIC.utils.cli import (
     get_analysis_fastq_files_directory,
 )
 from BALSAMIC.utils.io import write_json
-from BALSAMIC.models.analysis import ConfigModel
 
 LOG = logging.getLogger(__name__)
 
 
 @click.command("case", short_help="Create a sample config file from input sample data")
-@click.option(
-    "--case-id",
-    required=True,
-    help="Sample id that is used for reporting, \
-              naming the analysis jobs, and analysis path",
-)
-@click.option(
-    "--gender",
-    required=False,
-    default="female",
-    show_default=True,
-    type=click.Choice([Gender.FEMALE, Gender.MALE]),
-    help="Case associated gender",
-)
-@click.option(
-    "--umi/--no-umi",
-    default=True,
-    show_default=True,
-    is_flag=True,
-    help=(
-        "UMI processing steps for samples with UMI tags. For WGS cases, UMI is always disabled."
-    ),
-)
-@click.option(
-    "--umi-trim-length",
-    default=5,
-    show_default=True,
-    type=int,
-    help="Trim N bases from reads in fastq",
-)
-@click.option(
-    "--quality-trim/--no-quality-trim",
-    default=True,
-    show_default=True,
-    is_flag=True,
-    help="Trim low quality reads in fastq",
-)
-@click.option(
-    "--adapter-trim/--no-adapter-trim",
-    default=True,
-    show_default=True,
-    is_flag=True,
-    help="Trim adapters from reads in fastq",
-)
-@click.option(
-    "--fastq-path",
-    type=click.Path(exists=True, resolve_path=True),
-    required=True,
-    help="Path to directory containing unconcatenated fastqs",
-)
-@click.option(
-    "-p",
-    "--panel-bed",
-    type=click.Path(exists=True, resolve_path=True),
-    required=False,
-    help="Panel bed file for variant calling.",
-)
-@click.option(
-    "-b",
-    "--background-variants",
-    type=click.Path(exists=True, resolve_path=True),
-    required=False,
-    help="Background set of valid variants for UMI",
-)
-@click.option(
-    "--pon-cnn",
-    type=click.Path(exists=True, resolve_path=True),
-    required=False,
-    help="Panel of normal reference (.cnn) for cnvkit",
-)
-@click.option(
-    "--balsamic-cache",
-    type=click.Path(exists=True, resolve_path=True),
-    required=True,
-    help="Path to BALSAMIC cache",
-)
-@click.option(
-    "--container-version",
-    show_default=True,
-    default=balsamic_version,
-    type=click.Choice(["develop", "master", balsamic_version]),
-    help="Container for BALSAMIC version to download",
-)
-@click.option(
-    "--analysis-dir",
-    type=click.Path(exists=True, resolve_path=True),
-    required=True,
-    help="Root analysis path to store analysis logs and results. \
-                                     The final path will be analysis-dir/sample-id",
-)
-@click.option("--tumor-sample-name", type=str, required=True, help="Tumor sample name")
-@click.option(
-    "--normal-sample-name", type=str, required=False, help="Normal sample name"
-)
-@click.option(
-    "--cadd-annotations",
-    type=click.Path(exists=True, resolve_path=True),
-    required=False,
-    help="Path of CADD annotations",
-)
-@click.option(
-    "--clinical-snv-observations",
-    type=click.Path(exists=True, resolve_path=True),
-    required=False,
-    help="VCF path of clinical SNV observations (WGS analysis workflow)",
-)
-@click.option(
-    "--clinical-sv-observations",
-    type=click.Path(exists=True, resolve_path=True),
-    required=False,
-    help="VCF path of clinical SV observations (WGS analysis workflow)",
-)
-@click.option(
-    "--cancer-germline-snv-observations",
-    type=click.Path(exists=True, resolve_path=True),
-    required=False,
-    help="VCF path of cancer germline SNV normal observations (WGS analysis workflow)",
-)
-@click.option(
-    "--cancer-somatic-snv-observations",
-    type=click.Path(exists=True, resolve_path=True),
-    required=False,
-    help="VCF path of cancer SNV tumor observations (WGS analysis workflow)",
-)
-@click.option(
-    "--cancer-somatic-sv-observations",
-    type=click.Path(exists=True, resolve_path=True),
-    required=False,
-    help="VCF path of cancer SV observations (WGS analysis workflow)",
-)
-@click.option(
-    "--swegen-snv",
-    type=click.Path(exists=True, resolve_path=True),
-    required=False,
-    help="VCF path of Swegen SNV frequency database (WGS analysis workflow)",
-)
-@click.option(
-    "--swegen-sv",
-    type=click.Path(exists=True, resolve_path=True),
-    required=False,
-    help="VCF path of Swegen SV frequency database (WGS analysis workflow)",
-)
-@click.option(
-    "-g",
-    "--genome-version",
-    default="hg19",
-    type=click.Choice(["hg19", "hg38", "canfam3"]),
-    help=(
-        "Genome version to prepare reference. Path to genome will be <outdir>/genome_version"
-    ),
-)
-@click.option(
-    "-w",
-    "--analysis-workflow",
-    default="balsamic",
-    show_default=True,
-    type=click.Choice(["balsamic", "balsamic-umi", "balsamic-qc"]),
-    help=(
-        'Analysis workflow to run. By default: "balsamic" only '
-        "workflow will be running. If you want to run both "
-        "balsamic and UMI workflow together for panel data; "
-        'choose "balsamic-umi" option '
-    ),
-)
+@OPTION_ADAPTER_TRIM
+@OPTION_ANALYSIS_DIR
+@OPTION_ANALYSIS_WORKFLOW
+@OPTION_BACKGROUND_VARIANTS
+@OPTION_BALSAMIC_CACHE
+@OPTION_CADD_ANNOTATIONS
+@OPTION_CANCER_GERMLINE_SNV_OBSERVATIONS
+@OPTION_CANCER_SOMATIC_SNV_OBSERVATIONS
+@OPTION_CANCER_SOMATIC_SV_OBSERVATIONS
+@OPTION_CASE_ID
+@OPTION_CLINICAL_SNV_OBSERVATIONS
+@OPTION_CLINICAL_SV_OBSERVATIONS
+@OPTION_CONTAINER_VERSION
+@OPTION_FASTQ_PATH
+@OPTION_GENDER
+@OPTION_GENOME_VERSION
+@OPTION_NORMAL_SAMPLE_NAME
+@OPTION_PANEL_BED
+@OPTION_PON_CNN
+@OPTION_QUALITY_TRIM
+@OPTION_SWEGEN_SNV
+@OPTION_SWEGEN_SV
+@OPTION_TUMOR_SAMPLE_NAME
+@OPTION_UMI
+@OPTION_UMI_TRIM_LENGTH
 @click.pass_context
 def case_config(
-    context,
-    case_id,
-    gender,
-    umi,
-    umi_trim_length,
-    adapter_trim,
-    quality_trim,
-    fastq_path,
-    panel_bed,
-    background_variants,
-    pon_cnn,
-    analysis_dir,
-    tumor_sample_name,
-    normal_sample_name,
-    cadd_annotations,
-    clinical_snv_observations,
-    clinical_sv_observations,
-    cancer_germline_snv_observations,
-    cancer_somatic_snv_observations,
-    cancer_somatic_sv_observations,
-    swegen_snv,
-    swegen_sv,
-    genome_version,
-    balsamic_cache,
-    container_version,
-    analysis_workflow,
+    context: click.Context,
+    adapter_trim: bool,
+    analysis_dir: Path,
+    analysis_workflow: AnalysisWorkflow,
+    background_variants: Path,
+    balsamic_cache: Path,
+    cadd_annotations: Path,
+    cancer_germline_snv_observations: Path,
+    cancer_somatic_snv_observations: Path,
+    cancer_somatic_sv_observations: Path,
+    case_id: str,
+    clinical_snv_observations: Path,
+    clinical_sv_observations: Path,
+    container_version: str,
+    fastq_path: Path,
+    gender: Gender,
+    genome_version: GenomeVersion,
+    normal_sample_name: str,
+    panel_bed: Path,
+    pon_cnn: Path,
+    quality_trim: bool,
+    swegen_snv: Path,
+    swegen_sv: Path,
+    tumor_sample_name: str,
+    umi: bool,
+    umi_trim_length: int,
 ):
     if container_version:
         balsamic_version = container_version