+++ BALSAMIC/commands/run/analysis.py
@@ -1,16 +1,40 @@
+"""Balsamic run analysis CLI."""
 import json
 import logging
 import os
 import subprocess
 import sys
 from pathlib import Path
+from typing import List
 
 import click
 
+from BALSAMIC.commands.options import (
+    OPTION_SNAKEFILE,
+    OPTION_SAMPLE_CONFIG,
+    OPTION_RUN_MODE,
+    OPTION_CLUSTER_PROFILE,
+    OPTION_RUN_ANALYSIS,
+    OPTION_CLUSTER_QOS,
+    OPTION_FORCE_ALL,
+    OPTION_SNAKEMAKE_OPT,
+    OPTION_CLUSTER_ACCOUNT,
+    OPTION_CLUSTER_MAIL,
+    OPTION_CLUSTER_MAIL_TYPE,
+    OPTION_QUIET,
+    OPTION_DISABLE_VARIANT_CALLER,
+    OPTION_BENCHMARK,
+    OPTION_CLUSTER_CONFIG,
+    OPTION_DRAGEN,
+)
+from BALSAMIC.constants.analysis import RunMode
+from BALSAMIC.constants.cluster import (
+    ClusterConfigType,
+    QOS,
+    ClusterMailType,
+    ClusterProfile,
+)
 from BALSAMIC.models.analysis import ConfigModel
-from BALSAMIC.constants.cluster import ClusterConfigType
-from BALSAMIC.constants.paths import SCHEDULER_PATH
-from BALSAMIC.constants.workflow_params import VCF_DICT
 from BALSAMIC.models.snakemake import SnakemakeExecutable
 from BALSAMIC.utils.analysis import get_singularity_bind_paths
 from BALSAMIC.utils.cli import (
@@ -20,169 +44,54 @@ from BALSAMIC.utils.cli import (
     get_config_path,
 )
 
-
 LOG = logging.getLogger(__name__)
 
 
-@click.command(
-    "analysis", short_help="Run the analysis on a provided sample config-file"
-)
-@click.option(
-    "-S",
-    "--snake-file",
-    type=click.Path(),
-    show_default=True,
-    help=(
-        "Input for a custom snakefile. WARNING: "
-        "This is for internal testing, and should "
-        "not be used. Providing a snakefile supersedes"
-        "analysis_type option."
-    ),
-)
-@click.option(
-    "-s",
-    "--sample-config",
-    required=True,
-    type=click.Path(),
-    help="Sample json config file.",
-)
-@click.option(
-    "--run-mode",
-    show_default=True,
-    default="cluster",
-    type=click.Choice(["local", "cluster"]),
-    help=(
-        "Run mode to use. By default SLURM will be used to "
-        "run the analysis. But local runner also available "
-        "for local computing"
-    ),
-)
-@click.option(
-    "-c",
-    "--cluster-config",
-    show_default=True,
-    default=get_config_path(ClusterConfigType.ANALYSIS),
-    type=click.Path(),
-    help="cluster config json file. (eg- SLURM, QSUB)",
-)
-@click.option(
-    "--dragen", is_flag=True, default=False, help="Enable dragen variant caller"
-)
-@click.option(
-    "-p",
-    "--profile",
-    default="slurm",
-    type=click.Choice(["slurm", "qsub"]),
-    help="cluster profile to submit jobs",
-)
-@click.option(
-    "--benchmark",
-    default=False,
-    is_flag=True,
-    help="Profile slurm jobs using the value of this option. Make sure you have slurm profiler enabled in your HPC.",
-)
-@click.option(
-    "-r",
-    "--run-analysis",
-    show_default=True,
-    default=False,
-    is_flag=True,
-    help=(
-        "By default balsamic run_analysis will run in "
-        "dry run mode. Raise thise flag to make the "
-        "actual analysis"
-    ),
-)
-@click.option(
-    "--qos",
-    type=click.Choice(["low", "normal", "high", "express"]),
-    show_default=True,
-    default="low",
-    help=f"QOS for sbatch jobs. Passed to {SCHEDULER_PATH.as_posix()}",
-)
-@click.option(
-    "-f",
-    "--force-all",
-    show_default=True,
-    default=False,
-    is_flag=True,
-    help="Force run all analysis. This is same as snakemake --forceall",
-)
-@click.option(
-    "--snakemake-opt", multiple=True, help="Pass these options directly to snakemake"
-)
-@click.option(
-    "--account",
-    "--slurm-account",
-    "--qsub-account",
-    help="cluster account to run jobs, ie: slurm_account",
-)
-@click.option(
-    "--mail-user", help="cluster mail user to send out email. e.g.: slurm_mail_user"
-)
-@click.option(
-    "-q",
-    "--quiet",
-    default=False,
-    is_flag=True,
-    help="Instruct snakemake to be quiet! No output will be printed",
-)
-@click.option(
-    "--mail-type",
-    type=click.Choice(
-        [
-            "NONE",
-            "BEGIN",
-            "END",
-            "FAIL",
-            "REQUEUE",
-            "ALL",
-            "TIME_LIMIT",
-        ]
-    ),
-    help=(
-        "cluster mail type to send out email. This will "
-        "be applied to all jobs and override snakemake settings."
-    ),
-)
-@click.option(
-    "--disable-variant-caller",
-    help=(
-        f"Run workflow with selected variant caller(s) disable."
-        f"Use comma to remove multiple variant callers. Valid "
-        f"values are: {list(VCF_DICT.keys())}"
-    ),
-)
+@click.command("analysis", short_help="Run the analysis on a sample config-file")
+@OPTION_BENCHMARK
+@OPTION_CLUSTER_ACCOUNT
+@OPTION_CLUSTER_CONFIG
+@OPTION_CLUSTER_MAIL
+@OPTION_CLUSTER_MAIL_TYPE
+@OPTION_CLUSTER_PROFILE
+@OPTION_CLUSTER_QOS
+@OPTION_DISABLE_VARIANT_CALLER
+@OPTION_DRAGEN
+@OPTION_FORCE_ALL
+@OPTION_QUIET
+@OPTION_RUN_ANALYSIS
+@OPTION_RUN_MODE
+@OPTION_SAMPLE_CONFIG
+@OPTION_SNAKEFILE
+@OPTION_SNAKEMAKE_OPT
 @click.pass_context
 def analysis(
-    context,
-    snake_file,
-    sample_config,
-    run_mode,
-    cluster_config,
-    run_analysis,
-    force_all,
-    snakemake_opt,
-    mail_type,
-    mail_user,
-    account,
-    qos,
-    profile,
-    disable_variant_caller,
-    quiet,
-    dragen,
-    benchmark,
+    context: click.Context,
+    snakefile: Path,
+    sample_config: Path,
+    run_mode: RunMode,
+    cluster_config: Path,
+    benchmark: bool,
+    dragen: bool,
+    profile: ClusterProfile,
+    run_analysis: bool,
+    qos: QOS,
+    force_all: bool,
+    snakemake_opt: List[str],
+    account: str,
+    mail_user: str,
+    mail_type: ClusterMailType,
+    quiet: bool,
+    disable_variant_caller: str,
 ):
-    """
-    Runs BALSAMIC workflow on the provided sample's config file
-    """
-    LOG.info(f"BALSAMIC started with log level {context.obj['loglevel']}.")
+    """Run BALSAMIC workflow on the provided sample's config file."""
+    LOG.info(f"BALSAMIC started with log level {context.obj['log_level']}.")
 
-    if run_mode == "cluster" and not run_analysis:
+    if run_mode == RunMode.CLUSTER and not run_analysis:
         LOG.info("Changing run-mode to local on dry-run")
-        run_mode = "local"
+        run_mode: RunMode = RunMode.LOCAL
 
-    if run_mode == "cluster" and not account:
+    if run_mode == RunMode.CLUSTER and not account:
         LOG.info(
             "slurm-account, qsub-account, or account is required for slurm run mode"
         )
@@ -222,7 +131,7 @@ def analysis(
     analysis_workflow = config_model.analysis.analysis_workflow
     analysis_dir: Path = Path(config_model.analysis.analysis_dir)
     snakefile: Path = (
-        snake_file if snake_file else get_snakefile(analysis_type, analysis_workflow)
+        snakefile if snakefile else get_snakefile(analysis_type, analysis_workflow)
     )
 
     LOG.info(f"Starting {analysis_workflow} workflow...")
@@ -230,7 +139,9 @@ def analysis(
         account=account,
         benchmark=benchmark,
         case_id=case_name,
-        cluster_config_path=cluster_config,
+        cluster_config_path=cluster_config
+        if cluster_config
+        else get_config_path(ClusterConfigType.ANALYSIS),
         config_path=sample_config_path,
         disable_variant_caller=disable_variant_caller,
         dragen=dragen,