+++ BALSAMIC/commands/run/base.py
@@ -1,13 +1,14 @@
+"""Balsamic run CLI."""
 import click
 
-from BALSAMIC.commands.run.analysis import analysis as run_analysis_cmd
+from BALSAMIC.commands.run.analysis import analysis as run_analysis_command
 
 
 @click.group()
 @click.pass_context
-def run(context):
-    "Run BALSAMIC on a provided config file"
+def run(context: click.Context):
+    """Run Balsamic analysis on a provided configuration file."""
     pass
 
 
-run.add_command(run_analysis_cmd)
+run.add_command(run_analysis_command)